font scripts combined
